+++ packages/digit_components/CHANGELOG.md
@@ -110,3 +110,6 @@
 
 ## 1.0.0+3
 * Capture and Hide Location Dialog added to utils 
+
+## 1.0.0+4
+* Added new component selected box